finished ramune
just kind of finished the rest of it. still could use some tweaking. not all of the neat effects are in here that i might've liked (a circular expanding click effect on the labels was originally in the plan) but i also didn't really want to overdo it. i've got pawbin things to do! page still runs fast in the up to date chrome. should work in firefox just as well. nice!