almost finished
everything is fully operation except for the enabled/disabled gallery navigation. the gallery thumbs are not as pleasant as i'd like them to be, but it seems to be impossible to have dynamically sized squares in a flexbox (you can only make one axis, width or height, adjust with flex). whatever! very close to having starfield feeling good again.